--- portal-web/test/com/liferay/portalweb/demo/fundamentals/knowledgebase/SearchKBAArticleContentKBSTest.java
@@ -1,56 +1,0 @@
-/**
- * Copyright (c) 2000-2012 Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portalweb.demo.fundamentals.knowledgebase;
-
-import com.liferay.portalweb.portal.BaseTestCase;
-import com.liferay.portalweb.portal.util.RuntimeVariables;
-
-/**
- * @author Brian Wing Shun Chan
- */
-public class SearchKBAArticleContentKBSTest extends BaseTestCase {
-	public void testSearchKBAArticleContentKBS() throws Exception {
-		selenium.open("/web/guest/home/");
-
-		for (int second = 0;; second++) {
-			if (second >= 90) {
-				fail("timeout");
-			}
-
-			try {
-				if (selenium.isVisible("link=Knowledge Base Search Test Page")) {
-					break;
-				}
-			}
-			catch (Exception e) {
-			}
-
-			Thread.sleep(1000);
-		}
-
-		selenium.clickAt("link=Knowledge Base Search Test Page",
-			RuntimeVariables.replace("Knowledge Base Search Test Page"));
-		selenium.waitForPageToLoad("30000");
-		selenium.type("//input[@id='_5_WAR_knowledgebaseportlet_keywords']",
-			RuntimeVariables.replace("content"));
-		selenium.clickAt("//input[@value='Search']",
-			RuntimeVariables.replace("Search"));
-		selenium.waitForPageToLoad("30000");
-		assertEquals(RuntimeVariables.replace("Knowledge Base Article 1"),
-			selenium.getText("//td[1]/a"));
-		assertEquals(RuntimeVariables.replace("Joe Bloggs"),
-			selenium.getText("//td[2]/a"));
-	}
-}
