+++ portal-web/test/com/liferay/portalweb/demo/knowledgebase/AddKBAArticleSectionsPortletsTest.html
