+++ portal-web/test/com/liferay/portalweb/demo/knowledgebase/ConfigurePortletKBSSectionsASTest.html
