+++ portal-web/test/com/liferay/portalweb/demo/knowledgebase/DeleteKBAArticleSectionsCMKBArTest.html
@@ -45,7 +45,12 @@
 </tr>
 <tr>
 	<td>assertText</td>
-	<td>//div[@class='portlet-msg-error']</td>
+	<td>xPath=(//div[@class='portlet-msg-error'])[1]</td>
+	<td>Your request failed to complete.</td>
+</tr>
+<tr>
+	<td>assertText</td>
+	<td>xPath=(//div[@class='portlet-msg-error'])[2]</td>
 	<td>The selected article no longer exists.</td>
 </tr>
 </tbody></table>
