+++ portal-web/test/com/liferay/portalweb/demo/knowledgebase/DeleteKBAArticleSectionsPortletsKBDTest.html
