+++ portal-web/test/com/liferay/portalweb/demo/knowledgebase/EditKBAArticleSectionsPortletsAttachmentTest.html
@@ -89,19 +89,19 @@
 	<td>Browse</td>
 </tr>
 <tr>
-	<td>type</td>
+	<td>uploadCommonFile</td>
 	<td>//input[@id='_1_WAR_knowledgebaseportlet_file']</td>
-	<td>L:\portal\build\portal-web\test\com\liferay\portalweb\demo\fundamentals\knowledgebase\dependencies\test_image.jpg</td>
+	<td>Document_1.jpg</td>
 </tr>
 <tr>
 	<td>waitForVisible</td>
 	<td>//a/span</td>
-	<td>test_image.jpg (12.9k)</td>
+	<td>Document_1.jpg (12.9k)</td>
 </tr>
 <tr>
 	<td>assertText</td>
 	<td>//a/span</td>
-	<td>test_image.jpg (12.9k)</td>
+	<td>Document_1.jpg (12.9k)</td>
 </tr>
 <tr>
 	<td>close</td>
@@ -121,12 +121,12 @@
 <tr>
 	<td>waitForVisible</td>
 	<td>//div[@id='_1_WAR_knowledgebaseportlet_attachments']/div/div/span/a/span</td>
-	<td>test_image.jpg (12.9k)</td>
+	<td>Document_1.jpg (12.9k)</td>
 </tr>
 <tr>
 	<td>assertText</td>
 	<td>//div[@id='_1_WAR_knowledgebaseportlet_attachments']/div/div/span/a/span</td>
-	<td>test_image.jpg (12.9k)</td>
+	<td>Document_1.jpg (12.9k)</td>
 </tr>
 <tr>
 	<td>clickAtAndWait</td>
