+++ portal-web/test/com/liferay/portalweb/demo/knowledgebase/KnowledgeBaseTests.html
@@ -8,14 +8,14 @@
 <body>
 <table id="suiteTable" cellpadding="1" cellspacing="1" border="1" class="selenium"><tbody>
 <tr><td><b>Test Suite</b></td></tr>
-<tr><td><a href="../../../../portalweb/plugins/knowledgebase/knowledgebasearticle/portlet/addportletkbar/AddPageKBArTest.html">AddPageKBArTest</a></td></tr>
-<tr><td><a href="../../../../portalweb/plugins/knowledgebase/knowledgebasearticle/portlet/addportletkbar/AddPortletKBArTest.html">AddPortletKBArTest</a></td></tr>
-<tr><td><a href="../../../../portalweb/plugins/knowledgebase/knowledgebasedisplay/portlet/addportletkbd/AddPageKBDTest.html">AddPageKBDTest</a></td></tr>
-<tr><td><a href="../../../../portalweb/plugins/knowledgebase/knowledgebasedisplay/portlet/addportletkbd/AddPortletKBDTest.html">AddPortletKBDTest</a></td></tr>
-<tr><td><a href="../../../../portalweb/plugins/knowledgebase/knowledgebasesearch/portlet/addportletkbsr/AddPageKBSrTest.html">AddPageKBSrTest</a></td></tr>
-<tr><td><a href="../../../../portalweb/plugins/knowledgebase/knowledgebasesearch/portlet/addportletkbsr/AddPortletKBSrTest.html">AddPortletKBSrTest</a></td></tr>
-<tr><td><a href="../../../../portalweb/plugins/knowledgebase/knowledgebasesection/portlet/addportletkbsc/AddPageKBScTest.html">AddPageKBScTest</a></td></tr>
-<tr><td><a href="../../../../portalweb/plugins/knowledgebase/knowledgebasesection/portlet/addportletkbsc/AddPortletKBScTest.html">AddPortletKBScTest</a></td></tr>
+<tr><td><a href="../../../portalweb/plugins/knowledgebase/knowledgebasearticle/portlet/addportletkbar/AddPageKBArTest.html">AddPageKBArTest</a></td></tr>
+<tr><td><a href="../../../portalweb/plugins/knowledgebase/knowledgebasearticle/portlet/addportletkbar/AddPortletKBArTest.html">AddPortletKBArTest</a></td></tr>
+<tr><td><a href="../../../portalweb/plugins/knowledgebase/knowledgebasedisplay/portlet/addportletkbd/AddPageKBDTest.html">AddPageKBDTest</a></td></tr>
+<tr><td><a href="../../../portalweb/plugins/knowledgebase/knowledgebasedisplay/portlet/addportletkbd/AddPortletKBDTest.html">AddPortletKBDTest</a></td></tr>
+<tr><td><a href="../../../portalweb/plugins/knowledgebase/knowledgebasesearch/portlet/addportletkbsr/AddPageKBSrTest.html">AddPageKBSrTest</a></td></tr>
+<tr><td><a href="../../../portalweb/plugins/knowledgebase/knowledgebasesearch/portlet/addportletkbsr/AddPortletKBSrTest.html">AddPortletKBSrTest</a></td></tr>
+<tr><td><a href="../../../portalweb/plugins/knowledgebase/knowledgebasesection/portlet/addportletkbsc/AddPageKBScTest.html">AddPageKBScTest</a></td></tr>
+<tr><td><a href="../../../portalweb/plugins/knowledgebase/knowledgebasesection/portlet/addportletkbsc/AddPortletKBScTest.html">AddPortletKBScTest</a></td></tr>
 <tr><td><a href="AddKBAArticleSectionsPortletsTest.html">AddKBAArticleSectionsPortletsTest</a></td></tr>
 <tr><td><a href="ViewKBAArticleSectionsPortletsTest.html">ViewKBAArticleSectionsPortletsTest</a></td></tr>
 <tr><td><a href="EditKBAArticleSectionsPortletsAttachmentTest.html">EditKBAArticleSectionsPortletsAttachmentTest</a></td></tr>
@@ -33,7 +33,7 @@
 <tr><td><a href="SearchKBAArticleTomcatKBSTest.html">SearchKBAArticleTomcatKBSTest</a></td></tr>
 <tr><td><a href="DeleteKBAArticleSectionsCMKBArTest.html">DeleteKBAArticleSectionsCMKBArTest</a></td></tr>
 <tr><td><a href="TearDownKBAArticleTest.html">TearDownKBAArticleTest</a></td></tr>
-<tr><td><a href="../../../../portalweb/portal/util/TearDownPageTest.html">TearDownPageTest</a></td></tr>
+<tr><td><a href="../../../portalweb/portal/util/TearDownPageTest.html">TearDownPageTest</a></td></tr>
 </tbody></table>
 </body>
 </html>
