+++ portal-web/test/com/liferay/portalweb/demo/knowledgebase/SearchKBAArticleContentKBSTest.html
