+++ portal-web/test/com/liferay/portalweb/demo/knowledgebase/SearchKBAArticleKnowledgeKBSTest.html
