+++ portal-web/test/com/liferay/portalweb/demo/knowledgebase/SearchKBAArticleTomcatKBSTest.html
