+++ portal-web/test/com/liferay/portalweb/demo/knowledgebase/SelectKBAArticleSectionsCMKBArTest.html
