+++ portal-web/test/com/liferay/portalweb/demo/knowledgebase/ViewConfigurePortletKBSSectionsASTest.html
