+++ portal-web/test/com/liferay/portalweb/demo/knowledgebase/ViewEditKBAArticleSectionsPortletsAttachmentTest.html
@@ -36,7 +36,7 @@
 <tr>
 	<td>assertText</td>
 	<td>//div[@class='kb-article-attachments']/div/span/a/span</td>
-	<td>test_image.jpg (12.9k)</td>
+	<td>Document_1.jpg (12.9k)</td>
 </tr>
 </tbody></table>
 </body>
