+++ portal-web/test/com/liferay/portalweb/demo/knowledgebase/ViewKBAArticleSectionsPortletsTest.html
