+++ portal-web/test/com/liferay/portalweb/demo/knowledgebase/ViewSelectKBAArticleSectionsCMKBArTest.html
