+++ portal-web/test/com/liferay/portalweb/portal/DemoKnowledgeBaseTestSuite.java
@@ -12,13 +12,10 @@
  * details.
  */
 
-package com.liferay.portalweb.demo.fundamentals;
+package com.liferay.portalweb.portal;
 
-import com.liferay.portalweb.demo.fundamentals.announcements.AnnouncementsTests;
-import com.liferay.portalweb.demo.fundamentals.announcementsdistribitionscope.AnnouncementsDistributionScopeTests;
-import com.liferay.portalweb.demo.fundamentals.socialequityblogs.SocialEquityBlogsTests;
-import com.liferay.portalweb.demo.fundamentals.wsrp.WSRPTests;
-import com.liferay.portalweb.portal.BaseTestSuite;
+import com.liferay.portalweb.demo.knowledgebase.KnowledgeBaseTests;
+import com.liferay.portalweb.portal.login.LoginTests;
 
 import junit.framework.Test;
 import junit.framework.TestSuite;
@@ -26,15 +23,15 @@ import junit.framework.TestSuite;
 /**
  * @author Brian Wing Shun Chan
  */
-public class FundamentalsTestPlan extends BaseTestSuite {
+public class DemoKnowledgeBaseTestSuite extends BaseTestSuite {
 
 	public static Test suite() {
 		TestSuite testSuite = new TestSuite();
 
-		testSuite.addTest(AnnouncementsTests.suite());
-		testSuite.addTest(AnnouncementsDistributionScopeTests.suite());
-		testSuite.addTest(SocialEquityBlogsTests.suite());
-		testSuite.addTest(WSRPTests.suite());
+		testSuite.addTest(LoginTests.suite());
+		testSuite.addTest(KnowledgeBaseTests.suite());
+
+		testSuite.addTestSuite(StopSeleniumTest.class);
 
 		return testSuite;
 	}
